Separate Sum impl
